pyproject.toml
[tool.poetry]
name = "home-assignment"
version = "0.1.0"
description = "Web Developer Challenge"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
strawberry-graphql = {version = "^0.253.1", extras = ["fastapi"]}
uvicorn = "^0.32.1"
beanie = "^1.28.0"
pydantic-settings = "^2.1.0"
pyjwt = "^2.10.1"
fastapi = "^0.115.6"

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.2"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120
[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
